+++ Runtime/Model/BacktraceConfiguration.cs
@@ -13,6 +13,21 @@ namespace Backtrace.Unity.Model
     [CreateAssetMenu(fileName = "Backtrace Configuration", menuName = "Backtrace/Configuration", order = 0)]
     public class BacktraceConfiguration : ScriptableObject
     {
+        private const BacktraceBreadcrumbType AllBreadcrumbsTypes =
+            BacktraceBreadcrumbType.Configuration |
+            BacktraceBreadcrumbType.Http |
+            BacktraceBreadcrumbType.Log |
+            BacktraceBreadcrumbType.Manual |
+            BacktraceBreadcrumbType.Navigation |
+            BacktraceBreadcrumbType.System |
+            BacktraceBreadcrumbType.User;
+
+        private const UnityEngineLogLevel AllLogTypes = UnityEngineLogLevel.Debug |
+            UnityEngineLogLevel.Error |
+            UnityEngineLogLevel.Fatal |
+            UnityEngineLogLevel.Info |
+            UnityEngineLogLevel.Warning;
+
         /// <summary>
         /// Backtrace server url
         /// </summary>
@@ -30,6 +45,12 @@ namespace Backtrace.Unity.Model
         /// </summary>
         [Tooltip("Reports per minute: Limits the number of reports the client will send per minutes. If set to 0, there is no limit. If set to a higher value and the value is reached, the client will not send any reports until the next minute. Default: 50")]
         public int ReportPerMin = 50;
+
+        /// <summary>
+        /// "Disable error reporting integration in editor mode.
+        /// </summary>
+        [Tooltip("Disable error reporting integration in editor mode.")]
+        public bool DisableInEditor = false;
 
         /// <summary>
         /// Determine if client should catch unhandled exceptions
@@ -50,12 +71,12 @@ namespace Backtrace.Unity.Model
         public bool DestroyOnLoad = false;
 
         /// <summary>
-        /// Sampling configuration - fractional sampling allows to drop some % of unhandled exception.
+        /// Sampling configuration - fractional sampling allows to drop some % of Unity errors.
         /// </summary>
-        [Tooltip("Log random sampling rate - Enables a random sampling mechanism for unhandled exceptions - by default sampling is equal to 0.01 - which means only 1% of randomply sampling reports will be send to Backtrace. \n" +
-            "* 1 - means 100% of unhandled exception reports will be reported by library,\n" +
-            "* 0.1 - means 10% of unhandled exception reports will be reported by library,\n" +
-            "* 0 - means library is going to drop all unhandled exception.")]
+        [Tooltip("Log random sampling rate - Enables a random sampling mechanism for Unity.Error logs - by default sampling is equal to 0.01 - which means only 1% of randomply sampling reports will be send to Backtrace. \n" +
+            "* 1 - means 100% of error reports will be reported by library,\n" +
+            "* 0.1 - means 10% of error reports will be reported by library,\n" +
+            "* 0 - means library is going to drop all errors.")]
         [Range(0, 1)]
         public double Sampling = 0.01d;
 
@@ -160,13 +181,13 @@ namespace Backtrace.Unity.Model
         /// Backtrace breadcrumbs log level controls what type of information will be available in the breadcrumbs file
         /// </summary>
         [Tooltip("Breadcrumbs support breadcrumbs level- Backtrace breadcrumbs log level controls what type of information will be available in the breadcrumb file")]
-        public BacktraceBreadcrumbType BacktraceBreadcrumbsLevel;
+        public BacktraceBreadcrumbType BacktraceBreadcrumbsLevel = AllBreadcrumbsTypes;
 
         /// <summary>
         /// Backtrace Unity Engine log Level controls what log types will be included in the final breadcrumbs file
         /// </summary>
         [Tooltip("Breadcrumbs log level")]
-        public UnityEngineLogLevel LogLevel;
+        public UnityEngineLogLevel LogLevel = AllLogTypes;
 
         /// <summary>
         /// Use normalized exception message instead environment stack trace, when exception doesn't have stack trace
@@ -196,7 +217,7 @@ namespace Backtrace.Unity.Model
         /// Directory path where reports and minidumps are stored
         /// </summary>
         [Tooltip("This is the path to directory where the Backtrace database will store reports on your game. NOTE: Backtrace database will remove all existing files on database start.")]
-        public string DatabasePath;
+        public string DatabasePath = "${Application.persistentDataPath}/backtrace";
 
         /// <summary>
         /// Enable event aggregation support
@@ -234,7 +255,7 @@ namespace Backtrace.Unity.Model
         /// Determine if BacktraceDatabase should try to create database directory on application start
         /// </summary>
         [Tooltip("If toggled, the library will create the offline database directory if the provided path doesn't exists.")]
-        public bool CreateDatabase = false;
+        public bool CreateDatabase = true;
 
         /// <summary>
         /// Maximum number of stored reports in Database. If value is equal to zero, then limit not exists
@@ -296,7 +317,7 @@ namespace Backtrace.Unity.Model
             {
                 int universeIndexStart = backtraceSubmitUrl.Length;
                 int universeIndexEnd = submissionUrl.IndexOf('/', universeIndexStart);
-                if(universeIndexEnd == -1)
+                if (universeIndexEnd == -1)
                 {
                     throw new ArgumentException("Invalid Backtrace URL");
                 }