+++ Runtime/Model/Breadcrumbs/IBacktraceBreadcrumbs.cs
@@ -7,6 +7,8 @@ namespace Backtrace.Unity.Model.Breadcrumbs
     public interface IBacktraceBreadcrumbs
     {
         BacktraceBreadcrumbType BreadcrumbsLevel { get; }
+        bool EnableBreadcrumbs();
+        [Obsolete("Please use EnableBreadcrumbs instead. This function will be removed in the future updates")]
         bool EnableBreadcrumbs(BacktraceBreadcrumbType level, UnityEngineLogLevel unityLogLevel);
         bool ClearBreadcrumbs();
         bool Log(string message, LogType type, IDictionary<string, string> attributes);