+++ Runtime/Native/iOS/NativeClient.cs
@@ -6,6 +6,8 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading;
 using Backtrace.Unity.Model;
+using Backtrace.Unity.Model.Breadcrumbs;
+using Backtrace.Unity.Runtime.Native.Base;
 using UnityEngine;
 
 namespace Backtrace.Unity.Runtime.Native.iOS
@@ -13,26 +15,8 @@ namespace Backtrace.Unity.Runtime.Native.iOS
     /// <summary>
     /// iOS native client 
     /// </summary>
-    internal class NativeClient : INativeClient
+    internal class NativeClient : NativeClientBase, INativeClient
     {
-        // Last Backtrace client update time 
-        volatile internal float _lastUpdateTime;
-
-        /// <summary>
-        /// Determine if the ANR background thread should be disabled or not 
-        /// for some period of time.
-        /// This option will be used by the native client implementation
-        /// once application goes to background/foreground
-        /// </summary>
-        volatile internal bool _preventAnr = false;
-
-        /// <summary>
-        /// Determine if ANR thread should exit
-        /// </summary>
-        volatile internal bool _stopAnr = false;
-
-        private Thread _anrThread;
-
         // NSDictinary entry used only for iOS native integration
         internal struct Entry
         {
@@ -55,6 +39,9 @@ namespace Backtrace.Unity.Runtime.Native.iOS
         [DllImport("__Internal", EntryPoint = "AddAttribute")]
         private static extern void AddAttribute(string key, string value);
 
+        [DllImport("__Internal", EntryPoint = "Disable")]
+        private static extern void DisableNativeIntegration();
+
         private static bool INITIALIZED = false;
 
         /// <summary>
@@ -68,22 +55,19 @@ namespace Backtrace.Unity.Runtime.Native.iOS
 
 #endif
 
-        public NativeClient(BacktraceConfiguration configuration, IDictionary<string, string> clientAttributes, ICollection<string> attachments)
+        public NativeClient(BacktraceConfiguration configuration, BacktraceBreadcrumbs breadcrumbs, IDictionary<string, string> clientAttributes, ICollection<string> attachments) : base(configuration, breadcrumbs)
         {
             if (INITIALIZED || !_enabled)
             {
                 return;
             }
-            if (configuration.CaptureNativeCrashes)
+            if (_configuration.CaptureNativeCrashes)
             {
-                HandleNativeCrashes(configuration, clientAttributes, attachments);
+                HandleNativeCrashes(clientAttributes, attachments);
                 INITIALIZED = true;
             }
-            if (configuration.HandleANR)
+            if (_configuration.HandleANR)
             {
-                // iOS integration doesn't require to pass game object name or callback function
-                // it's required by android to know which one object to call when ANR was detected 
-                // in Java. In this situation we simply ignore them.
                 HandleAnr();
             }
         }
@@ -93,9 +77,9 @@ namespace Backtrace.Unity.Runtime.Native.iOS
         /// Start crashpad process to handle native Android crashes
         /// </summary>
 
-        private void HandleNativeCrashes(BacktraceConfiguration configuration, IDictionary<string, string> attributes, IEnumerable<string> attachments)
+        private void HandleNativeCrashes(IDictionary<string, string> attributes, IEnumerable<string> attachments)
         {
-            var databasePath = configuration.GetFullDatabasePath();
+            var databasePath = _configuration.GetFullDatabasePath();
             // make sure database is enabled 
             if (string.IsNullOrEmpty(databasePath) || !Directory.Exists(databasePath))
             {
@@ -103,16 +87,17 @@ namespace Backtrace.Unity.Runtime.Native.iOS
                 return;
             }
 
-            var plcrashreporterUrl = new BacktraceCredentials(configuration.GetValidServerUrl()).GetPlCrashReporterSubmissionUrl();
+            var plcrashreporterUrl = new BacktraceCredentials(_configuration.GetValidServerUrl()).GetPlCrashReporterSubmissionUrl();
 
             // add exception.type attribute to PLCrashReporter reports
             // The library will send PLCrashReporter crashes to Backtrace
             // only when Crash occured
-            attributes["error.type"] = "Crash";
+            attributes[ErrorTypeAttribute] = CrashType;
             var attributeKeys = attributes.Keys.ToArray();
             var attributeValues = attributes.Values.ToArray();
 
-            Start(plcrashreporterUrl.ToString(), attributeKeys, attributeValues, attributeValues.Length, configuration.OomReports, attachments.ToArray(), attachments.Count());
+            Start(plcrashreporterUrl.ToString(), attributeKeys, attributeValues, attributeValues.Length, _configuration.OomReports, attachments.ToArray(), attachments.Count());
+            CaptureNativeCrashes = true;
         }
 
         /// <summary>
@@ -154,7 +139,7 @@ namespace Backtrace.Unity.Runtime.Native.iOS
         /// <summary>
         /// Setup iOS ANR support and set callback function when ANR happened.
         /// </summary>
-        public void HandleAnr(string gameObjectName = "", string callbackName = "")
+        public void HandleAnr()
         {
             // if INITIALIZED is equal to false, plcrashreporter instance is disabled
             // so we can't generate native report
@@ -165,26 +150,27 @@ namespace Backtrace.Unity.Runtime.Native.iOS
 
             bool reported = false;
             var mainThreadId = Thread.CurrentThread.ManagedThreadId;
-            _anrThread = new Thread(() =>
+            AnrThread = new Thread(() =>
             {
                 float lastUpdatedCache = 0;
-                while (_anrThread.IsAlive && _stopAnr == false)
+                while (AnrThread.IsAlive && StopAnr == false)
                 {
-                    if (!_preventAnr)
+                    if (!PreventAnr)
                     {
                         if (lastUpdatedCache == 0)
                         {
-                            lastUpdatedCache = _lastUpdateTime;
+                            lastUpdatedCache = LastUpdateTime;
                         }
-                        else if (lastUpdatedCache == _lastUpdateTime)
+                        else if (lastUpdatedCache == LastUpdateTime)
                         {
                             if (!reported)
                             {
+                                OnAnrDetection();
                                 // set temporary attribute to "Hang"
-                                SetAttribute("error.type", "Hang");
-                                NativeReport("ANRException: Blocked thread detected.", true, true);
+                                SetAttribute(ErrorTypeAttribute, HangType);
+                                NativeReport(AnrMessage, true, true);
                                 // update error.type attribute in case when crash happen 
-                                SetAttribute("error.type", "Crash");
+                                SetAttribute(ErrorTypeAttribute, CrashType);
                                 reported = true;
                             }
                         }
@@ -194,7 +180,7 @@ namespace Backtrace.Unity.Runtime.Native.iOS
                         }
 
 
-                        lastUpdatedCache = _lastUpdateTime;
+                        lastUpdatedCache = LastUpdateTime;
                     }
                     else if (lastUpdatedCache != 0)
                     {
@@ -206,8 +192,8 @@ namespace Backtrace.Unity.Runtime.Native.iOS
 
                 }
             });
-            _anrThread.IsBackground = true;
-            _anrThread.Start();
+            AnrThread.IsBackground = true;
+            AnrThread.Start();
         }
 
 
@@ -250,34 +236,17 @@ namespace Backtrace.Unity.Runtime.Native.iOS
             // native plugin will analyse previous application session             
             return true;
         }
-
-        /// <summary>
-        /// Update native client internal timer.
-        /// </summary>
-        /// <param name="time">Current time</param>
-        public void UpdateClientTime(float time)
-        {
-            _lastUpdateTime = time;
-        }
-
         /// <summary>
         /// Disable native client integration
         /// </summary>
-        public void Disable()
+        public override void Disable()
         {
-            if (_anrThread != null)
+            if (CaptureNativeCrashes)
             {
-                _stopAnr = true;
+                CaptureNativeCrashes = false;
+                DisableNativeIntegration();
             }
-        }
-
-        /// <summary>
-        /// Pause ANR detection
-        /// </summary>
-        /// <param name="stopAnr">True - if native client should pause ANR detection"</param>
-        public void PauseAnrThread(bool stopAnr)
-        {
-            _preventAnr = stopAnr;
+            base.Disable();
         }
     }
 }