+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Backtrace Unity Release Notes
 
+## Version 3.7.0
+
+- When an ANR/Hang is detected, it is now added to the Breadcrumbs on all the platforms we support ANRs for
+- Sampling has been enabled in the editor mode.
+- Disable API available in the editor mode. API allows to ignore all reports generated in the Unity editor.
+- Disable NDK integration when Application.Exit API has been invoked.
+- Java background exception handler
+- BacktraceConfiguration default options change for DatabasePath and CreateDatabaseDirectory properties.
+- Updated Backtrace-Android dependencies to version 3.6.0
+
 ## Version 3.6.2
 
 Bugfixes